export now causes network side-effects, the tests are moved to integration tests
